Update chocolateyinstall.ps1
Please always use download.pdf24.org because download.pdf24.org distributes the load to different servers. Currently there are 5 servers (downloadX.pdf24.org) which run in parallel and each of them is limited in bandwidth. Currently we see too much traffic on download2.pdf24.org. So, please use download.pdf24.org, if this is possible and follow the redirect.